requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Deprecated==1.2.15
fastapi==0.115.6
filelock==3.13.1
fonttools==4.55.3
fsspec==2024.2.0
grpcio==1.69.0
h11==0.14.0
h5py==3.12.1
idna==3.10
Jinja2==3.1.3
kiwisolver==1.4.8
limits==4.0.1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.2.1
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.3
pydantic==2.10.5
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
scipy==1.15.1
setuptools==70.0.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
torch==2.5.1+cu124
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
wrapt==1.17.2
